refactor: extract LSP methods from app.go into app_lsp.go
Move all LSP-related App methods (autocomplete, signature help,
formatting, rename, references, definitions, text edits, document
notifications) into app_lsp.go (~840 lines). Rename lsp.go to
lsp_convert.go to clarify its role as a type conversion bridge.

app.go drops from 1281 to 457 lines, retaining only the App struct,
sidebar/panel/terminal management, hover, status messages, and
dialog helpers.
